mmc-utils: lsmmc.c: print_mmc_cid(): correct year/month parsing
The parsing logic for the MMC manufacturing date had the year and month
swapped.  From JESD84:

The manufacturing date, MDT, is composed of two hexadecimal digits, four
bits each, representing a two digits date code m/y; The “m” field, most
significant nibble, is the month code.  1 = January.  The “y” field, least
significant nibble, is the year code.

Notice that this is the opposite of the SD bit ordering.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
Reviewed-by: Shawn Lin <shawn.lin@rock-chips.com>
Signed-off-by: Avri Altman <avri.altman@sandisk.com>
Acked-by: Peter Korsgaard <peter@korsgaard.com>
Reviewed-by: Shawn Lin <shawn.lin@rock-chips.com>
Link: https://lore.kernel.org/r/20251104200008.940057-1-peter@korsgaard.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>